# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Garage API"
  description: "API for managing an auto garage, including customers, inventory, mechanics, and service tickets."
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
definitions:
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: "Resource not found."
  ValidationErrorResponse:
    type: object
    properties:
      messages:
        type: object
        additionalProperties:
          type: array
          items:
            type: string
        example:
          email: ["Invalid email format.", "Email already exists."]
          password: ["Field may not be null."]
...
  InventoryItemSummary: # For nesting
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      name:
        type: string
        example: "Spark Plug"
      price:
        type: string
        example: "15.99"

  # Service Ticket Definitions
  ServiceTicketCore:
    type: object
    properties:
      vin:
        type: string
        example: "123ABC456DEF789GHI"
      service_date:
        type: string
        format: date
        example: "2024-07-26"
      service_desc:
        type: string
        example: "Oil change and tire rotation"
      customer_id:
        type: integer
        format: int64
        example: 1
    required:
      - vin
      - service_date
      - customer_id
  CreateServiceTicketPayload:
    allOf:
      - $ref: "#/definitions/ServiceTicketCore"
  ServiceTicketResponse:
    allOf:
      - $ref: "#/definitions/ServiceTicketCore"
      - type: object
        properties:
          id:
            type: integer
            format: int64
            readOnly: true
            example: 1
          customer:
            $ref: "#/definitions/CustomerSummaryForTicket"
            readOnly: true
          mechanics:
            type: array
            items:
              $ref: "#/definitions/MechanicSummaryForTicket"
            readOnly: true
          inventory:
            type: array
            items:
              $ref: "#/definitions/ServiceTicketInventoryItemResponse"
            readOnly: true
...
  CustomerSummaryForTicket: # For nesting in ServiceTicket
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      name:
        type: string
        example: "John Doe"
      email:
        type: string
        format: email
        example: "john.doe@example.com"
      phone:
        type: string
        example: "555-1234"
  MechanicSummaryForTicket: # For nesting in ServiceTicket
    type: object
    properties:
      id:
        type: integer
        format: int64
        example: 1
      name:
        type: string
        example: "Jane Smith"
      email:
        type: string
        format: email
        example: "jane.smith@example.com"
      phone:
        type: string
        example: "555-5678"
      salary:
        type: string
        example: "60000.00"
  ServiceTicketInventoryItemBase:
    type: object
    properties:
      inventory_id:
        type: integer
        format: int64
        example: 1
      quantity:
        type: integer
        example: 2
    required:
      - inventory_id
      - quantity
  ServiceTicketInventoryItemResponse: # For ServiceTicket.inventory list
    allOf:
      - $ref: "#/definitions/ServiceTicketInventoryItemBase"
      - type: object
        properties:
          id: # This is ServiceInventory.id
            type: integer
            format: int64
            readOnly: true
            example: 101 
          inventory:
            $ref: "#/definitions/InventoryItemSummary"
            readOnly: true
...
  EditServiceTicketMechanicsPayload:
    type: object
    properties:
      add_mechanic_ids:
        type: array
        items:
          type: integer
          format: int64
        example: [1, 2]
      remove_mechanic_ids:
        type: array
        items:
          type: integer
          format: int64
        example: [3]
    required:
      - add_mechanic_ids
      - remove_mechanic_ids
  EditServiceTicketInventoryPayload:
    type: object
    properties:
      items_to_add_or_update:
        type: array
        items:
          $ref: "#/definitions/ServiceTicketInventoryItemBase"
      remove_inventory_ids:
        type: array
        items:
          type: integer
          format: int64
        example: [10, 12]
  MessageWithTicketResponse:
    type: object
    properties:
      message:
        type: string
      ticket:
        $ref: "#/definitions/ServiceTicketResponse"
paths:
...
  /service_tickets/:
    post:
      tags:
        - Service Tickets
      summary: "Create a new service ticket"
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateServiceTicketPayload"
          examples:
            application/json:
              vin: "NEWVIN123456789"
              service_date: "2024-08-01"
              service_desc: "Check engine light diagnosis"
              customer_id: 1
      responses:
        "201":
          description: "Service ticket created successfully"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
          examples:
            application/json:
              id: 104
              vin: "NEWVIN123456789"
              service_date: "2024-08-01"
              service_desc: "Check engine light diagnosis"
              customer_id: 1
              customer: { id: 1, name: "John Doe", email: "john.doe@example.com", phone: "555-1234" }
              mechanics: []
              inventory: []
        "400":
          description: "Validation error or VIN already exists"
          schema:
            oneOf:
              - $ref: "#/definitions/ErrorResponse"
              - $ref: "#/definitions/ValidationErrorResponse"
          examples:
            application/json_vin_exists:
              error: "Service ticket with VIN NEWVIN123456789 already exists"
            application/json_validation:
              messages:
                customer_id: ["Missing data for required field."]
    get:
      tags:
        - Service Tickets
      summary: "Get all service tickets"
      responses:
        "200":
          description: "A list of all service tickets"
          schema:
            type: array
            items:
              $ref: "#/definitions/ServiceTicketResponse"
          examples:
            application/json:
              - id: 101
                vin: "VIN123XYZ"
                service_date: "2024-07-28"
                service_desc: "Annual Checkup"
                customer_id: 1
                customer: { id: 1, name: "John Doe", email: "john.doe@example.com", phone: "555-1234" }
                mechanics: [ { id: 1, name: "Mechanic Mike", email: "mike@example.com", phone: "555-0001", salary: "70000.00" } ]
                inventory: [ { inventory_id: 1, quantity: 1, id: 201, inventory: { id: 1, name: "Oil Filter", price: "25.00" } } ]
              - id: 104
                vin: "NEWVIN123456789"
                service_date: "2024-08-01"
                service_desc: "Check engine light diagnosis"
                customer_id: 1
                customer: { id: 1, name: "John Doe", email: "john.doe@example.com", phone: "555-1234" }
                mechanics: []
                inventory: []
  /service_tickets/{ticket_id}/assign-mechanic/{mechanic_id}:
    put:
      tags:
        - Service Tickets
      summary: "Assign a mechanic to a service ticket"
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
          format: int64
        - name: mechanic_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: "Mechanic assigned successfully"
          schema:
            $ref: "#/definitions/MessageWithTicketResponse"
          examples:
            application/json:
              message: "Mechanic Jane Smith (ID: 1) assigned to Ticket ID: 104"
              ticket:
                id: 104
                vin: "NEWVIN123456789"
                service_date: "2024-08-01"
                service_desc: "Check engine light diagnosis"
                customer_id: 1
                customer: { id: 1, name: "John Doe", email: "john.doe@example.com", phone: "555-1234" }
                mechanics: [ { id: 1, name: "Jane Smith", email: "jane.smith@example.com", phone: "555-5678", salary: "60000.00" } ]
                inventory: []
        "404":
          description: "Ticket or Mechanic not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json_ticket_not_found:
              error: "Ticket not found."
            application/json_mechanic_not_found:
              error: "Mechanic not found."
  /service_tickets/{ticket_id}/remove-mechanic/{mechanic_id}:
    put:
      tags:
        - Service Tickets
      summary: "Remove a mechanic from a service ticket"
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
          format: int64
        - name: mechanic_id
          in: path
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: "Mechanic removed successfully"
          schema:
            $ref: "#/definitions/MessageWithTicketResponse"
          examples:
            application/json:
              message: "Mechanic Jane Smith (ID: 1) removed from Ticket ID: 104" # Corrected message based on typical remove operations
              ticket:
                id: 104
                vin: "NEWVIN123456789"
                service_date: "2024-08-01"
                service_desc: "Check engine light diagnosis"
                customer_id: 1
                customer: { id: 1, name: "John Doe", email: "john.doe@example.com", phone: "555-1234" }
                mechanics: [] # Mechanic is now removed
                inventory: []
        "404":
          description: "Ticket or Mechanic not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json_ticket_not_found:
              error: "Ticket not found."
            application/json_mechanic_not_found:
              error: "Mechanic not found."
  /service_tickets/{ticket_id}/edit/:
    put:
      tags:
        - Service Tickets
      summary: "Edit mechanics assigned to a service ticket"
      description: "Add or remove mechanics from a service ticket using their IDs."
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/EditServiceTicketMechanicsPayload"
          examples:
            application/json:
              add_mechanic_ids: [2] # Add Mechanic Bob
              remove_mechanic_ids: [1] # Remove Jane Smith (if she was previously assigned)
      responses:
        "200":
          description: "Mechanics updated successfully for the ticket"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
          examples:
            application/json:
              id: 104
              vin: "NEWVIN123456789"
              service_date: "2024-08-01"
              service_desc: "Check engine light diagnosis"
              customer_id: 1
              customer: { id: 1, name: "John Doe", email: "john.doe@example.com", phone: "555-1234" }
              mechanics: [ { id: 2, name: "Mechanic Bob", email: "bob.builder@example.com", phone: "555-1122", salary: "65000.00" } ]
              inventory: []
        "400":
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
          examples:
            application/json:
              messages:
                add_mechanic_ids: ["Not a valid list of integers."]
        "404": # Though route doesn't explicitly return 404 for ticket, it's good practice
          description: "Service ticket not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json: # Assuming a generic error if ticket not found during processing
              error: "An error occurred while updating mechanics."
  /service_tickets/{ticket_id}/edit_inventory/:
    put:
      tags:
        - Service Tickets
      summary: "Edit inventory items for a service ticket"
      description: "Add, update, or remove inventory items associated with a service ticket."
      parameters:
        - name: ticket_id
          in: path
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/EditServiceTicketInventoryPayload"
          examples:
            application/json:
              items_to_add_or_update: [
                { "inventory_id": 1, "quantity": 2 }, # Add/Update Oil Filter
                { "inventory_id": 2, "quantity": 4 }  # Add/Update Air Filter
              ]
              remove_inventory_ids: [3] # Remove an item with inventory_id 3 if present
      responses:
        "200":
          description: "Inventory items updated successfully for the ticket"
          schema:
            $ref: "#/definitions/ServiceTicketResponse"
          examples:
            application/json:
              id: 104
              vin: "NEWVIN123456789"
              service_date: "2024-08-01"
              service_desc: "Check engine light diagnosis"
              customer_id: 1
              customer: { id: 1, name: "John Doe", email: "john.doe@example.com", phone: "555-1234" }
              mechanics: [ { id: 2, name: "Mechanic Bob", email: "bob.builder@example.com", phone: "555-1122", salary: "65000.00" } ]
              inventory: [
                { inventory_id: 1, quantity: 2, id: 301, inventory: { id: 1, name: "Oil Filter", price: "25.00" } },
                { inventory_id: 2, quantity: 4, id: 302, inventory: { id: 2, name: "Air Filter", price: "22.50" } }
              ]
        "400":
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
          examples:
            application/json:
              messages:
                items_to_add_or_update: { "0": { "quantity": ["Quantity must be greater than 0."] } }
        "404":
          description: "Service ticket or inventory item not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
          examples:
            application/json_ticket_not_found:
              error: "Service ticket not found"
            application/json_item_not_found: # If an inventory item in the payload doesn't exist
              error: "Inventory item with ID 999 not found." # Example, actual message might vary